requirements.txt
google-api-python-client==2.166.0
google-auth-oauthlib==1.2.1
PyYAML==6.0.2